move time parsers to `time` package
