+++ src/parsers/mod.rs
@@ -1,3 +1,3 @@
 pub mod entry;
-pub mod parser;
+pub mod time;
 pub mod search;
